requirements.txt
absl-py==2.0.0
accelerate==0.17.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
attrs==23.1.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
coloredlogs==15.0.1
contextlib2==21.6.0
contourpy==1.2.0
cycler==0.12.1
daemoniker==0.2.3
decorator==4.4.2
diffusers==0.17.1
dill==0.3.7
dm-tree==0.1.8
docker-pycreds==0.4.0
einops==0.7.0
entmax==1.1
exceptiongroup==1.2.0
farama-notifications==0.0.4
filelock==3.13.1
fonttools==4.46.0
fsspec==2023.12.1
gitdb==4.0.11
gitpython==3.1.41
glcontext==2.5.0
glfw==1.12.0
google-auth==2.25.1
google-auth-oauthlib==0.4.6
grpcio==1.59.3
gym==0.21.0
gym3==0.3.3
gymnasium==0.29.1
huggingface-hub==0.19.4
humanfriendly==10.0
hydra-core==1.3.2
idna==3.6
imageio==2.33.0
imageio-ffmpeg==0.3.0
importlib-metadata==4.13.0
importlib-resources==6.1.1
inflect==6.0.4
iniconfig==2.0.0
jinja2==3.1.2
jsonlines==4.0.0
kiwisolver==1.4.5
kornia==0.7.0
lxml==4.9.3
markdown==3.5.1
markupsafe==2.1.3
matplotlib==3.8.2
mineclip==0.1
minedojo==0.1
ml-collections==0.1.1
moderngl==5.9.0
moviepy==1.0.3
mpmath==1.3.0
multiprocess==0.70.15
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
onnx
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.3
pillow==10.1.0
pluggy==1.3.0
praw==7.7.1
prawcore==2.4.0
proglog==0.1.10
protobuf
psutil==5.9.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.9
pyparsing==3.1.1
pyro4==4.82
pytest==7.4.3
python-dateutil==2.8.2
pytorch-msssim==1.0.0
pytz==2023.3.post1
pyyaml==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel-yaml==0.17.4
ruamel-yaml-clib==0.2.8
safetensors==0.4.1
sb3-contrib==1.6.2
scipy==1.8.0
sentry-sdk==1.39.2
serpent==1.41
setproctitle==1.3.3
shimmy==1.3.0
six==1.16.0
smmap==5.0.1
stable-baselines3==1.6.2
sympy==1.12
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
timm==0.9.16
tokenizers==0.13.3
tomli==2.0.1
torch==2.1.1
torch-tb-profiler==0.4.3
torchsummary==1.5.1
torchvision==0.16.1
tqdm==4.66.1
transformers==4.30.2
triton==2.1.0
typing-extensions==4.8.0
tzdata==2023.3
update-checker==0.18.0
urllib3==2.1.0
wandb==0.16.2
websocket-client==1.7.0
werkzeug==3.0.1
wget==3.2
x-transformers==0.27.1
xmltodict==0.13.0
zipp==3.17.0